+++ packages/carbon-graphs/src/js/controls/PairedResult/PairedResult.js
@@ -278,7 +278,7 @@ class PairedResult extends GraphContent {
     const drawBox = (boxPath) => {
       drawSelectionIndicator(graph.scale, graph.config, boxPath);
       drawLine(graph.scale, graph.config, boxPath);
-      drawPoints(graph.scale, graph.config, boxPath);
+      drawPoints(graph.scale, graph.config, boxPath, graph.legendSVG);
     };
     const types = ['high', 'mid', 'low'];
     types.forEach((type) => {
@@ -344,7 +344,7 @@ class PairedResult extends GraphContent {
      */
   redraw(graph) {
     clear(graph.svg, this.dataTarget);
-    draw(graph.scale, graph.config, graph.svg, this.dataTarget);
+    draw(graph.scale, graph.config, graph.svg, this.dataTarget, graph.legendSVG);
     return this;
   }
 }