+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## Unreleased
 
+* Changed
+  * Replaced `document` in all draw helpers with `graph.legendSVG` for better control.
+
 -   Migrated to the terra-graphs mono-rep.